Use autounboxing instead of explicit calls to `intValue()`, etc.
RELNOTES=n/a
PiperOrigin-RevId: 885069082

+++ android/guava/src/com/google/common/collect/ImmutableRangeSet.java
@@ -607,7 +607,7 @@ public final class ImmutableRangeSet<C extends Comparable> extends AbstractRange
         }
         result = size = Ints.saturatedCast(total);
       }
-      return result.intValue();
+      return result;
     }
 
     @Override
